+++ docs/CHANGE_LOG.md
@@ -16,6 +16,7 @@
 - Added new public function "removeLanguage()", which will remove a specific language that can be rendered by name.
 - Added new public function "getAllLanguages()", which will return the object that contains all the language details.
 - Added new public function "highlightElement()", which will render a specific DOM element (accepts either the element, or the elements ID).
+- Added new public function "getLanguage()", which will return the details for a specific language (by name).
 
 #### **Custom Triggers:**
 - Renamed the custom trigger "onRender" to "onRenderComplete".